+++ src/lsp/constants.ts
@@ -1,0 +1,6 @@
+import { DocumentSelector } from 'vscode-languageclient';
+
+export const SORBET_DOCUMENT_SELECTOR: DocumentSelector = [
+  { language: 'ruby', scheme: 'file' },
+  { language: 'ruby', scheme: 'sorbet' },
+];