+++ src/lsp/highlightType.ts
@@ -1,0 +1,9 @@
+/**
+ * Highlight type.
+ * See https://sorbet.org/docs/highlight-untyped
+ */
+export const enum HighlightType {
+  Disabled = 'disabled',
+  Everywhere = 'everywhere',
+  EverywhereButTests = 'everywhere-but-tests',
+}