+++ src/lsp/languageClient.ts
@@ -1,0 +1,72 @@
+import { ErrorHandler, InitializationFailedHandler } from 'vscode-languageclient';
+import { LanguageClient, ServerOptions } from 'vscode-languageclient/node';
+import { SORBET_DOCUMENT_SELECTOR } from './constants';
+import { InitializationOptions } from './initializationOptions';
+import { ReadFileRequest } from './readFileRequest';
+import { ShowOperationNotification } from './showOperationNotification';
+import { ShowSymbolRequest } from './showSymbolRequest';
+import { SorbetExtensionContext } from '../sorbetExtensionContext';
+import { WorkspaceDidChangeConfigurationNotification } from './workspaceDidChangeConfigurationNotification';
+
+export type SorbetClient = LanguageClient
+  & ReadFileRequest
+  & ShowOperationNotification
+  & ShowSymbolRequest
+  & WorkspaceDidChangeConfigurationNotification;
+
+/**
+ * Create a {@link LanguageClient client} for Sorbet.
+ */
+export function createClient(
+  context: SorbetExtensionContext,
+  serverOptions: ServerOptions,
+  errorHandler: ErrorHandler,
+): SorbetClient {
+  const client = new class extends LanguageClient {
+    error(message: string, data?: any, showNotification?: boolean | 'force')
+      : void {
+      // Override `force` to prevent notifications dialogs from showing up in
+      // unintended scenarios (still ocurring in LanguageClient v9).
+      // Example messages:
+      // - Sorbet client: couldn't create connection to server.
+      // - Connection to server got closed. Server will not be restarted.
+      super.error(
+        message,
+        data,
+        showNotification === 'force' ? true : showNotification,
+      );
+    }
+  }
+    (
+      'ruby',
+      'Sorbet',
+      serverOptions,
+      {
+        documentSelector: SORBET_DOCUMENT_SELECTOR,
+        errorHandler,
+        initializationFailedHandler: createInitializationFailedHandler(),
+        initializationOptions: createInitializationOptions(),
+        outputChannel: context.logOutputChannel,
+      });
+
+  return client;
+
+  function createInitializationFailedHandler(): InitializationFailedHandler {
+    const { log } = context;
+    return (error) => {
+      log.error('Failed to initialize Sorbet', error);
+      return false;
+    };
+  }
+
+  function createInitializationOptions()
+    : InitializationOptions {
+    const { configuration } = context;
+    return {
+      enableTypedFalseCompletionNudges: configuration.typedFalseCompletionNudges,
+      highlightUntyped: configuration.highlightUntyped,
+      supportsOperationNotifications: true,
+      supportsSorbetURIs: true,
+    };
+  }
+}