+++ src/lsp/showSymbolRequest.ts
@@ -1,0 +1,14 @@
+import {
+  CancellationToken, SymbolInformation, TextDocumentPositionParams,
+} from 'vscode-languageclient';
+
+/**
+ * See https://sorbet.org/docs/lsp#sorbetshowsymbol-request
+ */
+export interface ShowSymbolRequest {
+  sendRequest(
+    method: 'sorbet/showSymbol',
+    param: TextDocumentPositionParams,
+    token?: CancellationToken):
+    Promise<SymbolInformation>
+}